ADFA-2685: Delete 2nd onboarding slide, move privacy disclosure to slide 3
Streamlines onboarding from three slides to two by dropping the informational
"Next page: permissions" slide. The Privacy & analytics disclosure dialog —
previously triggered on slide 2's onSlideSelected — now fires on the
Permissions slide via the same listener hook, preserving the once-per-install
preference key so existing users aren't re-prompted.

Removes the now-unused fragment, layout, adapter, model, test helper, and
the five permissions_info_* strings (en + id).
