+++ app/src/main/java/com/itsaky/androidide/activities/OnboardingActivity.kt
@@ -43,7 +43,6 @@ import com.itsaky.androidide.app.configuration.IJdkDistributionProvider
 import com.itsaky.androidide.fragments.onboarding.GreetingFragment
 import com.itsaky.androidide.fragments.onboarding.OnboardingInfoFragment
 import com.itsaky.androidide.fragments.onboarding.PermissionsFragment
-import com.itsaky.androidide.fragments.onboarding.PermissionsInfoFragment
 import com.itsaky.androidide.models.JdkDistribution
 import com.itsaky.androidide.preferences.internal.prefManager
 import com.itsaky.androidide.tasks.doAsyncWithProgress
@@ -113,7 +112,6 @@ class OnboardingActivity : AppIntro2() {
         pulseAnimation = AnimationUtils.loadAnimation(this, R.anim.pulse_animation)
 
         addSlide(GreetingFragment())
-		addSlide(PermissionsInfoFragment())
 
 		if (!PackageUtils.isCurrentUserThePrimaryUser(this)) {
 			val errorMessage =
