--- app/src/main/java/com/itsaky/androidide/models/PermissionInfoItem.kt
@@ -1,8 +1,0 @@
-package com.itsaky.androidide.models
-
-import androidx.annotation.StringRes
-
-data class PermissionInfoItem(
-  @StringRes
-  val text: Int
-)
