--- app/src/main/res/layout/layout_permission_info_item.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<com.google.android.material.textview.MaterialTextView xmlns:android="http://schemas.android.com/apk/res/android"
-  android:id="@+id/text"
-  android:layout_width="match_parent"
-  android:layout_height="wrap_content"
-  android:layout_marginTop="12dp"
-  android:textAppearance="@style/TextAppearance.Material3.BodyMedium" />
