+++ resources/src/main/res/values/strings.xml
@@ -725,11 +725,6 @@
 	<string name="permission_desc_notifications">Allow Code on the Go to display notifications.</string>
 	<string name="permission_title_accessibility">Accessibility</string>
 	<string name="permission_desc_accessibility">Allow Code on the Go to access accessibility services for debugging.</string>
-	<string name="permissions_info_intro">Next page: permissions</string>
-	<string name="permissions_info_notifications">• <b>Notifications</b> (to receive status messages) On Android 12 and earlier, notifications are enabled already.</string>
-	<string name="permissions_info_storage">• <b>Storage</b> (to create project directories and files)</string>
-	<string name="permissions_info_install">• <b>Install packages</b> (to install the apps that you build)</string>
-	<string name="permissions_info_overlay_accessibility">• <b>Overlay window</b> (to use the debugger) If you can\'t enable this permission, go to <b>Settings \u2192 Apps \u2192 Code on the Go</b>. In the upper right Options menu, tap <b>Enable restricted settings</b>. Return to the installer and assign the permission. (The steps may vary on different devices.)</string>
 	<string name="finish_installation">Finish installation</string>
 	<string name="onboarding_title_permissions">Permissions</string>
 	<string name="onboarding_subtitle_permissions">Code on the Go requires the following permissions:</string>
